requirement.txt
langchain
langchain-core
langchain-community
langgraph

langchain-openai
langchain-google-genai
langchain-groq

python-dotenv
faiss-cpu
tiktoken

pypdf
unstructured
beautifulsoup4
lxml

fastapi
uvicorn
streamlit